+++ android/app/src/main/java/com/artgallery/map/Types.kt
@@ -1,6 +1,6 @@
 package com.artgallery.map
 
 class Geoloc (
-    var lat: Int,
-    var lng: Int,
+    var lat: Float,
+    var lng: Float,
 )
